+++ doc/CHANGELOG.md
@@ -1,5 +1,11 @@
 ### Changelog
 
+## 2.6.0
+* Use SheetJS to export selected specobjects as xlsx files
+* Removed .csv export of selected specobjects.
+* Added dialog with drag-and-drop to configure fields exported to xlsx files
+* Make "exclude rejected" a setting
+
 ## 2.5.2
 * Add support quoted search terms in VQL. Enclose whole term in either pairs of `"` or `'` characters.
   This permits spaces in search terms, without rewriting spaces as `\s`.